requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arxiv==3.0.0
astrapy==2.2.1
asttokens==3.0.1
async-timeout==4.0.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.136.1
feedparser==6.0.12
filelock==3.29.0
firecrawl-py==4.25.2
frozenlist==1.8.0
fsspec==2026.4.0
gitdb==4.0.12
GitPython==3.1.50
greenlet==3.5.0
groq==0.37.1
h11==0.16.0
h2==4.3.0
hf-xet==1.5.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.14.0
hyperframe==6.1.0
idna==3.13
ipython==8.39.0
itsdangerous==2.2.0
jedi==0.20.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.18
langchain-classic==1.0.7
langchain-community==0.4.1
langchain-core==1.3.3
langchain-groq==1.1.2
langchain-huggingface==1.2.2
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.3
markdown-it-py==4.2.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.21.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
parso==0.8.7
pillow==12.2.0
prompt_toolkit==3.0.52
propcache==0.5.2
protobuf==7.34.1
pure_eval==0.2.3
pyarrow==24.0.0
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
pydeck==0.9.2
Pygments==2.20.0
pymongo==4.17.0
PyMuPDF==1.27.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
pytz==2026.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.4.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
streamlit==1.57.0
sympy==1.14.0
tavily-python==0.7.24
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tqdm==4.67.3
traitlets==5.15.0
transformers==5.8.0
typer==0.25.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
uuid6==2025.0.1
uuid_utils==0.14.1
uvicorn==0.46.0
watchdog==6.0.0
wcwidth==0.7.0
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
